Adds an "fsnative_traces" key to the proc dictionaries.
fsnative_traces is roughly equivalent to fsaverage_traces, but it is
not aligned to a flatmap. Instead, the traces are split up into 500
equally spaced points each and the original contours are aligned to
their fsaverage-aligned versions; then the 500-point traces are
aligned to the fsaverage using that alignment matrix and translation
vector. This makes fsnative_traces similar to fsaverage_traces in
that they are maximally aligned but using a rigid alignment.